requirements.txt
##########
# DJANGO #
##########
django
djangorestframework
django-cors-headers

#########
# MONGO #
#########
djongo
dnspython

#########
# OTHER #
#########
python-dotenv
pytz